requirements.txt
Flask
yt_dlp
numpy
librosa
openai-whisper
